- Implemented function to calculate quantity for weighted item. - refactored code to allow for decimal quantities (where weight does not match weighted price for product).

+++ ShoppingBasket/Models/BasketProduct.cs
@@ -9,9 +9,6 @@ namespace ShoppingBasket.Models
     public class BasketProduct
     {
         public Product Product { get; set; }
-        public int Quantity { get; set; }
-
-
-
+        public decimal Quantity { get; set; }
     }
 }
